#526 Preparing unit tests for generic async method issue
